pyproject.toml
[tool.poetry]
name = "newton-ai"
version = "0.1.0"
description = "Multi-agent Tutor AI for all your academic needs!"
authors = ["Abhay Shukla <abhayshuklaistheone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-adk = "^1.0.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
